pyproject.toml
[tool.poetry]
name = "simpleblog"
version = "0.1.0"
description = "Simple Blog on Python Django"
authors = ["Farm Tonita <farmtointa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
psycopg2 = "^2.9.5"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
